general: drop support for API version 1.0
API version 1.0 is considered insecure and should therefore no longer be used.
We included support for it in 6.3.0 to support updating the server side first,
but future versions will no longer include support for 1.0.

Agent mTLS connections can still be manually disabled, but this is highly
discouraged for normal installations. More details:
https://github.com/keylime/keylime/security/advisories/GHSA-2m39-75g9-ff5r

Signed-off-by: Thore Sommer <mail@thson.de>