elchecking: error if policy name is invalid, change default to reject-all
The old behavior was that if the policy name was not found in the registry
the "accept-all" policy was chosen. This might make the user believe that
their setup was correctly because measured boot still works.

The behavior is changed to error on startup if the policy was misconfigured
and now instead of "accept-all" "reject-all" is used to have a more secure
default.

Signed-off-by: Thore Sommer <mail@thson.de>