signing: move exception handing to verify_signature()
We had error handling in verify_signature_from_file(), which was the
only point where we called verify_signature(), however a recent allowlist
processing refactoring added new calls to verify_signature(), and these
calls do not properly handle the possible exception it might throw.

Instead of adding error handling to each call of verify_signature(), let
us instead catch/log the possible exceptions there, and then we can count
on it always returning a boolean to indicate whether it succeeded or not.

Signed-off-by: Sergio Correia <scorreia@redhat.com>