ima_ast: Make entry parsing stricter
This changes str.split and str.strip invocations to avoid any implicit
treatment of the space (" "), which is used as the delimiter character
for measurement list entries.  Also simplify the ima-sig template
parsing by assuming that the "n-ng" part doesn't contain any spaces
and the delimiter character is always present even if the "sig" part
is empty.

Signed-off-by: Daiki Ueno <dueno@redhat.com>