Adds a per-agent counter for "successfull attestations" on Keylime.
This counter, "attestation_count" is now available via API calls (e.g.,
`keylime_tenant -c status`) and it is stored on the database to preserve
state in case the `verifier` is restarted. Given the availability of
this information, the old "first_verified" attribute is now removed.
This PR is the first one on a series of PRs to implement "Durable
Attestation", also known as "Offline Attestation" (enhancement #73).

Signed-off-by: Marcio Silva <marcio.a.silva@ibm.com>