Add script to convert old config to new config
The script extracts the values from the given configuration file and
replaces the fields in the templates.  The templates for the
configuration files in the new split format are also added.

The templates to convert the old single configuration file to the new
split format are added with the convertion script, as well as the
'adjust.py' script to adjust values that changed from the old
configuration option to the new format.

The directories that contains the templates, mapping, and 'adjust.py'
script for each version are expected to be in the 'templates' directory.
The versions are expected to be in the format 'MAJOR.MINOR', where
'MAJOR' and 'MINOR' are positive integers (e.g.: '12.34').

When the input configuration does not have a version, it is assumed to
be the old configuration format, and the minimum version number ('1.0')
is used.

The transformation for each version found in the 'templates' folder is
applied in crescent order.  It is possible to specify a target version
through the '--version' option.  The script stops applying
transformations when the specified target version is reached.

The new configuration file is described in:
https://github.com/keylime/enhancements/blob/master/72_config_and_simplify_tls.md

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>