This commit solves #1091 by adding a per-agent verifier-issued epoch timestamp
A new DB column, containing an epoch timestamp (from the verifier)
records when a quote was received. This new column is then exposed as an
additional parameter on the API (whenever agent status is requested).

In case a verifier thread dealing with an  agent dies (while the
verifier process remains responsive) or in case of a multi-verifier
configuration (where a whole verifier could die, while one keeps
requesting status from a different verifier), there is a way for an user
to make a decision about the "age" of the last successful attestation.

Signed-off-by: Marcio Silva <marcio.a.silva@ibm.com>