WIP: Implements enhancement #73 "Durable (Offline) Attestion"
This commit is just an "early preview" of the code we are already using
currently in our internal PoCs. Nothing will be done until after the
"configuration file split" is merged (@ansasaki).

The code is at a point where @lkatalin and @lukehinds can take a look,
as requested. At this moment, I have included two example backends for
persistent store - redis and file (with sqlalchemy coming soon) - and
one example backend for the transparency log, rekor (this one is pretty
primitive at the moment, it uses the `rekor-cli` executable until the
sigstore python code - https://github.com/sigstore/sigstore-python - is
mature enough for use)

I have also included a (crude) CLI (`keylime_attest`) which will be
further refined, as discused with @mpeters and others.

One of the problems that I intend to tackle soon is how to "cold
initialize" `ima_policy` from "outside" the `verifier` (any help from
@mbestravos is appreciated on this). Until then I am storing all the
attestation artifacts on the persistent store serialized with `pickle`
(instead of encoded with `json`)

Signed-off-by: Marcio Silva <marcio.a.silva@ibm.com>