A fix for "per-agent verifier-issued epoch timestamp"
In some database backends, such as MySQL, the use of a float as column
lead to representation and rounding errors. In order to preserve the
information in the most backend-agnostic way possible, we will simply
use integer (nobody really cares about fractions of seconds for the
purpose of fault detection/"lagging behind").

Signed-off-by: Marcio Silva <marcio.a.silva@ibm.com>