Do not install keylime nor configuration files during tests
Use a virtual environment for the python components and temporary
directories for the generated files.  The dependencies are still
installed in the system.

The used configuration files are generated in a temporary directory with
the changes necessary to run the tests. The installed configuration
files are not modified.

The files generated when running run_tests.sh will be removed at the end
of the execution unless the environment variable KEEP_TMP_FILES=1.

If the tests are executed using the run_tests.sh script, all the
necessary dependencies are installed and environment variables set.

When running test_restful.py outside run_tests.sh with RUST_TEST=1, it
expects the KEYLIME_RUST_AGENT and KEYLIME_RUST_CONF environment
variables set, respectively, with the path to the rust agent binary and
the rust agent configuration file.

If the environment variables are not set, it tries to find the
rust agent binary and configuration in an adjacent directory (i.e. in
the same location run_tests.sh downloads it).

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>