da: This commit implements most of the changes for #73 "Durable (Offline) Attestation".
With this commit in place, attestation artifacts are OPTIONALLY recorded
into a Persistent Store (time-series-like database), if the correct
configuration options are set (documentation will done in a subsequent
commit). By the default, nothing will happen and keylime will continue
to operate as it has always been.

This is not the whole story, and a subsequent PR will include a CLI
utility `keylime_attest` which can query the persistent store, retrieve
all data and verify the attestation status on all artifacts stored (using
the unmodified verifier code).

As discussed with many others, I have included three "example backends"
for the persistent store: "flat file", "redis" (but of course!) and
"sqlalchemy".

In addition to that, I have included an OPTIONAL backend for a
Transparency Log (for now, Rekor is the only one that I am aware of), as
well as a backend for a Time Stamp Authority (TSA), compatible with the
protocol RFC3161.

The rekor backend still uses the rekor-cli executable, and will do so
until the sigstore python code -
https://github.com/sigstore/sigstore-python - is mature enough for
"general use" (which they admit is still not the case)

As we also discussed at some length, the infrastructure required for a
it (i.e., a Persistent Store + Transparency Log + Time Stamp Authority)
is not conducive to a simple test, and for the moment none is included.
We will most certainly revisit the issue in the future.

For the time being, this feature will be marked as "experimental" (with
the accompaning log message making this fact explicit)

Signed-off-by: Marcio Silva <marcio.a.silva@ibm.com>