Move EK cert verification to cert_utils
We had partially addressed the issue we encountered when parsing some
certificates with python-cryptography by writing cert_utils module that
provided a single helper to parse the pubkey from a certificate.

However, we still were doing the EK validation in tpm_main using
python-cryptography, which means we would fall into the same parsing
problem again, since during the validation it would read all the certs
in the tpm cert store to check if it is signing the presented EK.

We address this issue by moving the EK cert verification to cert_utils:
we use the same approach as before, i.e. we parse the cert with pyasn1
when python-cryptography fails, and then use python-cryptography to do
the actual signature verification, as before.

By moving the method to cert_utils, it also becomes simpler to test it,
so in this commit we more tests to verify the methods work as expected.

Additionally, the updated EK verification is also capable of handling
ECDSA signatures, so this commit also fixes an issue that had been
previously reported.

Hovewer, malformed certs that are not parsed by python-cryptography
still pose an issue as python-cryptography will be unable to verify
their signatures. One possible way to deal with those would be to use
different tools (e.g. openssl) to do the signature verification, with
the `ek_check_script' option.

Signed-off-by: Sergio Correia <scorreia@redhat.com>