Move the execution of external EK check script to cert_utils
Keylime allows to specify an optional script to check the EK and/or EK
certtificate against an allowlist or any other additional EK processing
you want to do, via the `ek_check_script' option in tenant.conf.

In this commit we move the code path that runs this external script to
cert_utils, so it becomes simpler to test, and we also provide a sample
`ek-openssl-verify' script that uses openssl to do the validation of
an EK certificate provided via the EK_CERT env var, against the
certificates in the TPM cert store.

Signed-off-by: Sergio Correia <scorreia@redhat.com>