Very limited code fixup/cleanup on keylime_tenant CLI
This PR starts the cleanup and fixup of several small issues on
`keylime_tenant` code. First of all, make sure all "jsonification" of
resquest's responses happen at the static method `_jsonify_response`
(renamed from `_print_json_response`) which can now optionally print the
response. Second make sure this new function is invoked as often as
possible throughout the CLI code, with obvious exceptions being
documented. Third, the old `do_cv` method is renamed to `do_cvadd`, in
order to improve the code's understandability and maintainability.

This by no means fixes all the problems and issues on `keylime_tenant`,
but it is an incremental step in the right direction.

Signed-off-by: Marcio Silva <marcio.a.silva@ibm.com>