ima: Accept x509 certificates if no Subject Key Identifier is available
Also accept x509 certificates if they do not have a Subject Key Identifier
(SKID). Previously only certificates with SKID were accepted because the
SKID is required for the key to be used by the kernel keyrings and
keylime would only accept a key that could possibly also have served for
IMA signature verification in the kernel.

Return None in case no keyidv2 could be determined either due to missing
extensions or missing SKID and warn the use about it.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>