cloud_verifier: This is the first PR to address the scalability problems uncovered in #1167, starting with `agent` status.
While requesting `agent` status, either individually or in bulk
(`bulkinfo`), there is simply no need for loading anything from the
`allowlists` table on the database, especialmente if the leght of the
allowlist for a given agent is non-trivial (e.g., tens of megabytes).

In addition to that, two attributes for the `agent` - `allowlist_len`
and `mb_refstate_len` - in addition of being of dubious utility, were
being calculated in an incorrect manner. These were dutifully removed.

A meaningful attribute for mesured boot reference states -
`has_mb_refstate` is already implemented on this PR.

In a subsquent PR we will find a way to introduce meaningful attribute
for IMA allowlists (i.e., `has_ima_allowlist`, or something to that
effect), **without** requiring a full load from the database table.

Signed-off-by: Marcio Silva <marcio.a.silva@ibm.com>