allowlist: Add an AllowlistDBProxy acting as a proxy for the allowlist table
Implement an AllowlistDBProxy that acts as a proxy and cache to the
VerifierAllowlist DB table by holding the VerifierAllowlist DB entries in
a map whose keys are the IDs of the allowlists. Whenever one needs a
VerfierAllowlist DB entry one queries a AllowlistDBProxy singleton for an
entry with that allowlist ID. If the entry is not currently cached in
the map it will be loaded from the DB and otherwise every few seconds the
currently cached entry will be checked for an update and the latest version
of the VerfierAllowlist object will be returned. Every once-in-while
(60 seconds) all entries that have expired will be purged in order to
clear up memory of either deleted or unreferenced VerifierAllowlist DB
entries.

Resolves: https://github.com/keylime/keylime/issues/1167
Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>