Do not use default values that need reading the config in methods
Following up from the recent refactoring that moved the EK validation
to cert_utils, in a few places were added default method values that
were reading the configuration files directly.

It was not such a great idea becasue it then made those config files as
required to even import the modules.

Example "from keylime import cert_utils" now also requires that the
tenant configuration be available for getting the path for the TPM
cert store.

Let's stop doing that.

Signed-off-by: Sergio Correia <scorreia@redhat.com>