verifier: Fix return value in validate_ima_policy_data()
Fix the return value in validate_ima_policy_data() to return an error
string in case of no agent_data that were provided. The single caller
of this function expects a single return value in form of an error
message that it then puts into a response.

This fixes a recently introduced bug in patch
"common: Remove redundant return parameter from validate_ima_policy_data"

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>