ima: Raise ValueError on unsupported key types
Raise a ValueError when the parsing of a certificate returns a keytype
other than an RSAPublicKey or an EllipticCurvePublicKey so that callers
do not use these keys by putting them into an IMA policy for example.
The key is currently rejected when the keyidv2 is calculated for one of
the supported key types.

Adjust all callers of get_pubkey() and get_pubkey_from_file() other than
the test cases where known valid keys are being parsed.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>