verifier: Calculates the checksum for the whole IMA policy on the verifier #1198
A previous PR (#1181) introduced `checksum` as a separate column on the
DB for the `allowlists` table. The initial implementation focused on the
calculating the checksum for the allowlist, piggybacking already
existing code on the `tenant`. However, this is not enough to detect
full changes on the "IMA policy" (which will cover also the "exclude
list"). This PR now ensures that a checksum is calculated for the whole
policy, on the `verifier`.

Signed-off-by: Marcio Silva <marcio.a.silva@ibm.com>