pyright: Fix a pyright issue in ca_impl_openssl
Filter-out unsupported private key types and raise a ValueError before
trying to sign with the private key. Apply a filter of all those keys
that could have possibly been used for signing. Any other type of key
would have already thrown an exception since it was not possible to
sign an x509 certificate with it.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>