codestyle: Do not assign function parameter a new value in function
Do not assign the function parameter 'output' a new value inside
the function if its input was 'None'. Instead, use a different
variable to later on not confuse pyright that will still assume
that output may be 'None' and complain that it can be 'None' when
it must not be 'None' in a function call, even if output was given
a value.

Type annotate the 'output' function parameter to Optional[List[str]]
to make pyright happy.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>