verifier: fix for IMA policy checksum calculation
The "allowlist" field on IMA policy (JSON) has, inside its "meta"
subfield, a "timestamp" key,value pair (generated at the `verifier`).
Due to this fact, whenever the `checksum` (over the whole policy) is
calculated (also) at the `verifier`, the end result is different values
for policies which are actually identical (minus the aforementioned
`timestamp`).

While it could be possible to simply remove this timestamp, in order to
prepare for future changes (such as PR #1147), this PR simply truncates
the actual value of the timestamp, as it has no relation to the contents
(e.g., "allowlist") of the IMA policy.

Signed-off-by: Marcio Silva <marcio.a.silva@ibm.com>