elchecking: add support for MeasuredBoot when SecureBoot is disabled
In some setups SecureBoot is disabled. MeasuredBoot should still work in that
case. Most GRUB implementations use the SHIM_LOCK protocol to verify
signatures when SecureBoot is enabled which adds an
EV_EFI_BOOT_SERVICES_APPLICATION event for the vmlinuz image. This does not
happen when SecureBoot is disabled.

This introduces the flag "has_secureboot" to the example policy, when this
option is set to false instead of "kernel_authcode_sha256"
"vmlinuz_plain_sha256" is used for verifying the kernel.
Note that those two options are mutually exclusive.

create_mb_refstate is updated to detect if SecureBoot is disabled and can
generate valid refstates for those new options.

Signed-off-by: Thore Sommer <mail@thson.de>