config: Check for available config upgrade on startup
Check for available configuration upgrades when starting the services.

The check is performed by comparing the latest available configuration
upgrade template version with the current component configuration
version.

When an minor configuration upgrade is available, a INFO level message
will be logged informing the available configuration upgrade and the
command to run to perform an upgrade (keylime_upgrade_config).

When a major configuration upgrade is available, a WARNING level message
will be logged.

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>