tenant: reliable and consistent add/delete operations (fixes #1158)
As discussed in issue #1158, there are real-world scenarios where "add"
and/or "delete" operations via `keylime_tenant` CLI causes it to return
before the data was written by `verifier` on the database. This causes
inconsistent "status" (also via `keylime_tenant`) for a given `agent`,
in case the CLI invoked within a small time window after "add"/"delete".

Given the fact the code for `keylime_tenant` was somewhat messy (e.g., a
fixed 12-step loop with 0.4 seconds between attempts on "delete"), this
new code uses the already existing "retry_interval",
"exponential_backoff", and "max_retries" attributes to implement an
ordered "status" polling on the `verifier` right after an "add" or
"delete" operation, only returning when the desired state and the actual
state effectivelly match (or produce an explicit error message in case
it does not happen after multiple retries)

Signed-off-by: Marcio Silva <marcio.a.silva@ibm.com>