verifier: proper support for listening on 0.0.0.0 (fixes #705)
This PR fixes the problem of having a `verifier` listening on 0.0.0.0 (a
case we aim to support, as pointed out by @mpeters) and then adding and
deleting (with specific CLI options) an `agent`.

The error stems from the fact that the "verifier_ip" column on the
database was populated from the value (0.0.0.0) found on the
configuration file (`/etc/keylime/verfier.conf`), which is wrong/useless
as a network endpoint. Given the need for an IP address whenever the
`tenant` wishes to contact the `verifier` (option -v/--cv on CLI) the
contents of the aforementioned column should reflect the value supplied
by the former, rather than what is read from the configuration file for
the latter. This ensures we will not end up with 0.0.0.0 on the
"verifier_ip" column.

Signed-off-by: Marcio Silva <marcio.a.silva@ibm.com>