config: improve support for (log-based) debugging
This PR attempts to increase the ability to use logs (which can be
ingested by an analytics stack, such as ELK) in order to debug startup.
To this end, all `print` statements in config.py are replaced by calls
to a (basic) logger (given we cannot use the keylime logger due to
circular dependency), and a debug level log message is added whenever a
configuration file option is overriden by an environment variable
(ability introduced with PR #1274).

In addition to the above, the actual invocation of external commands
(e.g., tpm2-tools) are also logged with debug statments.

Signed-off-by: Marcio Silva <marcio.a.silva@ibm.com>