Add timestamp of last successful attestation to verifier API
This commit adds a last_successful_attestation column to the database to
record the timestamp at which a given agent was last determined to have
passed attestation. This value is updated to the current time according
to the verifier whenever the verifier receives a valid quote from the
agent and is exposed through the verifier's REST API. If no valid
attestation has been received for the lifetime of the agent, the value
is set to 0.

In addition to amending the API documentation to include the new
last_successful_attestation field, descriptions have also been added for
attestation_count (added in 5ed5d6d9) and last_received_quote (bb0c3e3).

It was considered whether last_successful_attestation could be inferred
from existing stored values and just exposed as a JSON field through the
API without a corresponding database column, but there is no way to do
so consistently across all possible agent states.

Note that when an existing Keylime installation is upgraded using the
included migration, last_successful_attestation is initially set to NULL.

Resolves: keylime/enhancements#74
See also: #1091

Signed-off-by: Jean Snyman <git@jsnyman.com>