db: switch to pessimistic disconnect handling
If sessions are idle for a long time, the SQL server might no longer have that
connection open. This causes the Keylime to error when its tries to query the
database.

This enabled pool_pre_ping which checks if the connection is still active and
uses a new connection if not.

More details can be found here:
https://docs.sqlalchemy.org/en/20/core/pooling.html#disconnect-handling-pessimistic

Signed-off-by: Thore Sommer <mail@thson.de>