create_policy: read the hashes from filelists-ext
The future release of createrepo_c will, optionally, provide a new
metadata, filelists-ext.xml, that will extend filelists.xml with the
hashes of the files listed in the RPM package.

During the remote repo analysis, if this file is available it will be
used as a shortcut instead of reading the RPM headers.

Signed-off-by: Alberto Planas <aplanas@suse.com>