End of term for @mpeters + propose @maugustosilva
This PR is for me to step down as PTL of the project and let
someone else have a chance to run as PTL for a term.

I will still 100% be active on the project.

I would also like to make a proposal of @maugustosilva taking over from
me. Marcio has been very active in the keylime community for several years
in developing new features, reviewing code and working with others on their
designs.

The process is outlined here:

https://github.com/keylime/keylime/blob/master/GOVERNANCE.md#changes-in-project-lead

Typically the PR should be opened no earlier then 6 weeks before
the end of term, however I'm willing to move sooner if the community agrees.

At the same time, anyone else is welcome to raise a PR with a
nomination of their own.

I will leave this PR open for 2 weeks (unless objections are made
and a community member feels this should be longer) unless there is serious
disagreement.

I will also mail the list with details of my stepping down.

Signed-off-by: Michael Peters <mpeters@redhat.com>