verifier: Read list of agents early on
Read the list of agents early on from the Db and pass this list towards
activate_agents() where it was previously read. Reading the list of
agents early on will allow us to create other lists of agents to activate
per verifier process (in the next step) and work with a stable list.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>