ca:CLI utilities (keylime_ca,keylime_tenant) read password from ca.conf
Fixes #1135

This is a minimalistic fix for the issue of keystore certificate
passords. Instead of prompting an user to enter a password during the
exection of commands such as `keylime_ca --command init` or
`keylime_tenant -c add ... --cert <cert dir>`, the new code now simply
reads it from `/etc/keylime/ca.conf`.

A delicate balance between the old and new behavior was achieved by
implementing the following balance: the "password" attribute can never
be empty (on ca.conf) but if the "default" value is found there, all the
CLI utilities (i.e., `keylime_tenant` and `keylime_ca`) will stop and ask
for password, while the "server" components (i.e., `verifier`) com
operate with "default" as the password, as it was before (emitting a
WARN in this case, same behavior as before).

Please note that an user can specify multiple different passwords (which
can be required in case certificates to multiple nodes are created) by
simply executing the commands prefaced by
`KEYLIME_CA_PASSWORD=<password>`

Signed-off-by: Marcio Silva <marcio.a.silva@ibm.com>