tenant: fixes a (timing) issue whenever an agent is removed and re-added
The following timing issue has been detected on the `agent` (thanks
@ansasaki for the help debugging): `agent` does not replace K until both
the new U and V are received. The `tenant` tries to verify (due to the
cli option `--verify`) that the `agent` has the new K before it receives
the V part. For this request, the `agent` will use the old K, resulting
in failure on `tenant` side.

This minimalistic fix implements a retry (actually it just re-uses the
retry already existing on the code), just to give the `verifier` enough
time to send the new `V`.

Fixes #1380

Signed-off-by: Marcio Silva <marcio.a.silva@ibm.com>