tenant: log cleanup and output improvements
This PR is one more which modifies the log messages (this time with the
introduction of a "full ID string") in order to make easier to observe
and debug Keylime with tools such as Elasticsearch/Kibana (e.g., by
using the `agent`'s UUID as a search term to aggregate all events). This
rewriting of log message resulted in the elimination of several poorly
written messages.

This PR also ensures the output of commands `bulkinfo`, `status`,
`cvstatus`, `regstatus`, `reglist`, `cvlist` is "consumable" by a tool
such as `jq` (something like "<command> | grep ^{ | jq ." will work)

Signed-off-by: Marcio Silva <marcio.a.silva@ibm.com>