Implement automatic agent API version bump
Automatically update the agent supported API version in the database if
the agent is updated and its API version is bumped.

Previously, if an agent was added to a verifier while it used an old API
version, and then it is updated with an API version bump, the
attestation would fail as the verifier would try to reach the agent
using the old API version.

Fixes #1457

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>