ima: Emit a warning when a file signature could not be parsed
CentOS currently has file signatures where the signature length indicator
seems to be written in little endian. A failure to parse the length
indicator causes Keylime to silently discard the signature and the result
then is that the pcr template hash is not calculated correctly (missing
signature) for the file and then the running hash becomes wrong and one
ends up with this error message in the log:

IMA measurement list does not match TPM PCR f5e0272e4....

Log the failure if a non zero-length signature was attempted to be
parsed.

Resolves: https://github.com/keylime/keylime/issues/1462

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>