templates: verifier: Add require_allow_list_signatures to config file
Have the require_allow_list_signatures config option appear in the
verifier config file for the first time.

To maintain backwards compatility and not change the usage of
keylime_tenant and suddenly require that policies be signed, set the
above config option to 'False' by default so that signed policies are
not required as before.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>