runtime policy: Raise error on unsigned policy if signature required
Raise an error if the runtime policy is not signed but the verifier config
option requires that the policy be signed. So far signatures were only
required when a 'payload' item was found in the map of the given runtime
policy, which is not the case at all when the policy doesn't have a
signature.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>