ima: Report an error if no quote forward-progress was made
Report an error if the TPM's quote was not making forward-progress,
which means that the PCR that was previously quoted was lagging
behind the end of the log (= was not quote of PCR state corresponding
to last entry) AND that the current quote has not caught up with
the entire log from last time.

Add a test case testing the check_quote_progress() function.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>