templates: Restore the default values
This change is to restore the default values to the same as before the
change on the --default option behavior for the convert_config.py
script.

Previously, some default values on the mapping were corrected by the
adjust.py script on the template version 2.0, and these adjusted values
were kept as the default value.

With the change in behavior of the --default option, the default value
from the latest template mapping is used instead. This caused some
default values to diverge from what was used before, because of the lack
of the adjust script.

With this fix, the default value from version 2.3 will be in the correct
form without the need for adjustments with the 'adjust.py' script

The following options default values in the mappings files were modified
by the adjust script from template version 2.0:

- tenant:
    - accept_tpm_hash_algs
    - accept_tpm_encryption_algs
    - accept_tpm_signing_algs
- verifier:
    - revocations:
        - enabled_revocation_notifications

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>