ima: Implement IMAPolicy and rules classes for support of new IMA policy
Implement an IMAPolicy class that can parse text-based rules and create
a list of instances representing these rules. Have it support an eval
method that passes a line from the IMA log and has the entries in the
list evaluate the line one after another until one returns a REJECT
along with a failure. If none of the rules evaluates the line then the
line will be REJECT'ed.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>