WIP: ima: Implement support for reject map and filter list
WIP:
 - filter list
 - how to 'manage' these maps? how to add to files to them etc.

Implement support for a reject map that uses filenames as key to the map
and where values are hashes. The reject map can be used to collect files
that must not be installed on the system.

To use the reject map a user must write an IMA policy now and place it
into the runtime policy. It is not part of the internal default policy.

Extend the create_policy tool so that the rejects map is copied from a
base policy. (WIP)

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>