ima: Implement support for LEARN-KEYS policy rule
Implement a LEARN-KEYS policy rule has the capability to learn the keys
used by the system by looking at IMA log lines in the ima-buf format.
The rule has the following format:

LEARN-KEYS [ignored-list-ref=<ignore-list>] [allowed-hashes-ref=<hashes-list>]

The rule can reference a list of keyrings to ignore. A list like this is
embedded in the runtime policy under the name 'ignored_keyrings'.

The rule can also reference a map with allowed hashes for keys to be loaded.
This map uses keyring names as keys.

Add this rule to the internal policy.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>