PSS padding fix - salt length changed to byte length of digest from length of signature
The salt length for PSS should be the byte length of the digest getting signed. This has only been working when it was the same
size as the signature.

For example, if the item is hashed into 32 bytes (sha256) and signed with that amount, the verification would pass as the salt would
receive the correct length (256/8 = 32). However if the digest is 48 bytes, and the signature is 512 then it will fail
(512/8 = 64) as a salt length of 64 is longer than 48.

Signed-off-by: Isaac-Matthews <isaac.matthews@hpe.com>