ima: Extend FILE-HASHES rules with support for filename and hash parameters
Add support for filename and hash parameters to the FILE-HASHES rule with
the follow format for these options:

FILE-HASHES: filename=<filename> hash=<hash> [hash=<hash> ...] target=<ACCEPT|REJECT>

This allows to write rules that select a specific filename and compare the
hash of the file against those provided in the rule.

Add a test case.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>