ima: Implement a DEBUG policy rule to display input and evaluations
Implement a DEBUG policy rule supporting the following format:

DEBUG: [mode=on|off]

This rule can turn debugging of rule evaluations on or off. When it is
turned on then parts of the IMA log line are displayed and all subsequently
evaluated IMApolicy rules and their evalation results on the input data are
also displayed until rule evaluation ends with a rule that accepted the IMA
log line.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>