ima: Add support for rejects to policy and evaluation
Add support for rejects to the runtime policy as an optional entry in the
map but extend the EMPTY_RUNTIME_POLICY with a empty rejects map. Extend
the evaluation of the policy to process the optional rejects before the
exclude list so that the exclude list does not hide the rejects entries.

Implement a function trim_runtime_policy that sets the minimum policy
version on a policy after removing for example an empty rejects map.
This allows users to use a policy that makes no use of new features on
older versions of keylime. Call this function wherever the
EMPTY_RUNTIME_POLICY is used to create a new policy and where the policy
is written later on, such as in tools that create a runtime policy.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>