cmd/keylime_policy: add tool to handle keylime policies
Initial steps of enhancement #110, which is to provide a single
tool that provides the functionality of the existing tools we
have.

In this commit, we add the keylime_policy tool, that is currently
able to handle the creation of runtime policies.

Some examples of usage:

To see the available options:

keylime_policy --help

To see the options available related to creating runtime policies:

keylime_policy create runtime --help

Some of the options available include e.g. using an IMA measurement
list, legacy allowlist/excludelist files, local/remote RPM repositories,
specifying a rootfs filesystem, initrds, a base policy, etc.

Signed-off-by: Sergio Correia <scorreia@redhat.com>