Add core functionality for agent-driven (push) attestation
This is commit 1/2 which consolidates the work of PR #1693 to
add a new attestation protocol which is initiated by the agent
instead of the verifier, as has been the case previously. The
basic two phase protocol is implemented herein, but the final
REST API, data structures and code organisation are finalialised
in the subsequent commit.

Additionally included are some changes to the registrar, but these
are reverted in the next commit. They are intended to be contributed
separately in PR #1670.

This works builds off of earlier commit f88f2cc to implement
keylime/enhancements#103.

Squashes commits: 3908cce, 8214121, c8b674f, a63c60c, 7cfc6ce,
1966d22, cf6f50b, 57649b4, 912145f, d986339, cb3016a
For context, refer to PR #1693.

Co-authored-by: Jean Snyman <jean.snyman@hpe.com>

Signed-off-by: gnurugs <supreshna.gurung@hpe.com>
Signed-off-by: Jean Snyman <jean.snyman@hpe.com>