keylime_policy: Calculate digests from each source separately
Previously, the digests would be obtained from each source
incrementally, modifying the dictionary from the base policy or, if not
provided, from an empty policy.

This changes makes the resulting dictionary to be calculated separately
and combined at the end.

These changes also make the IMA measurement list to only be accessed
when requested with --use-ima-measurement-list

Finally, these changes makes the tool to check the digest algorithm used
in each part and will log warnings when different algorithms are used in
different parts.

The algorithm used when calculating digests (i.e. when measuring files
from a given rootfs or ramdisk) will be obtained from one of the
following sources, with this priority order:

1. --algo option, if provided
2. Base policy, if provided
3. Allowlist, if provided
4. IMA measurement list, if provided

If none of the previous was provided, the default digest algorithm will
be used. Currently, the default algorithm is SHA-1.

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>