tee/snp: Modify parameters, verify report freshness
A nonce is required to verify the freshness of SEV-SNP attestation
reports.

It cannot be reasonably expected that a guest would know which
SEV microarchitecture it is running under. Therefore, remove the
required "gen" argument. Default to Milan for now, but eventually cycle
through Milan, Genoa, and Turin.

Along with the nonce hash, the TEE public key's components (x and y EC
coordinates) are hashed into the REPORT_DATA. This is to ensure
freshness of the attestation reports.

Signed-off-by: Tyler Fanelli <tfanelli@redhat.com>