mba: normalize vendor_db in EV_EFI_VARIABLE_AUTHORITY events
tpm2_eventlog may provide the vendor_db data as either a parsed signature
list or raw hex bytes, depending on the version used.

In this commit we add a enrich_vendor_db_authority_variable() function to
make sure we end up with a signature list independent on the format of
the data obtained from tpm2_eventlog.

Signed-off-by: Sergio Correia <scorreia@redhat.com>