verify/evidence: Use tee label for TEE verification
Rather than specify the tee in the type argument for the
/verify/evidence handler, specify `tee` in the type argument and
redirect to a TEE handler. This handler will parse the resulting JSON
data accordingly and forward the evidence to the respective TEE
attestation handler.

This allows for easier specification of other evidence types as well as
specifying multiple evidence types in the future (for example, CPU + GPU
attestation in one request).

Signed-off-by: Tyler Fanelli <tfanelli@redhat.com>